Measure memory directly from the allocator

+++ macaco-base/src/algorithm.rs
@@ -14,7 +14,7 @@ pub trait Algorithm<T: Distance + Clone> {
     /// (second element)
     fn time_profile(&self) -> (Duration, Duration);
 
-    fn memory_usage(&self) -> Option<u64>;
+    fn memory_usage(&self) -> Option<usize>;
 
     fn counters(&self) -> (u64, u64);
 }
